This where the user goes away. REOVED: user and user_session - including models, views, controllers, routes and generators (which I had just recently added). ADDED: A cucumber/aruba "test" that will create a rails application, install blacklight, and verify that files are corrected added and or modified.   At the end it copies over the tests so that can run them in a fresh instance if you so desire.  (Also added Aruba - an extension to cucumber that allows you to easily execute command line arguments). CHANGED: blacklight_controller - changed the methods current_user_session and current_user to assume there is a method user_session and current_user available to it.  This assumption should be spelled out the documentation. blacklight_helper - for some reason facet links where including duplicate facet parameters - I modified one of the methods in blacklight_helper to make certain these were not duplicated - there is a deeper problem here, where parameters are getting set when they shouldn't - like all possible facets being set when you view a specific document.  I'm still trying to track this problem down. user_util_links -  changed logout and login paths to destroy_user_session and new_user_session - which allows us to work with devise out of the box.  These are simply assumed to be restful routes defined in the hosting applications routes. Comma_link_renderer was moved into lib/blacklight/  which makes it far easier to locate when loading all of this as a Gem.  Also include will_paginate in blacklight.rb - which causes it to load at the right time. blacklight_engine - now causes ActiveRecord base to include /lib/blacklight/user.rb - which makes the is_blacklight_user method available. Removed the /user/authlogic.rb file.  And moved the contents of /user_generated_content into the user object - just to keep things simple - you now have one place to look for methods injected into active record objects that call is_blacklight_user Generator:  Blacklight generator now take a -d or --devise argument which will case it to add devise to your Gemfile, run bundle install, and run three devise generators:  devise:install, devise user, and devise:views Cha MIGRATIONS:  Migrations now exist as a part of the generator - I did not want to add a migration that removes the user table - this seems like a really bad idea - the alternative is not to call the migration anymore - so there is no user migration.  Also - added a user_type field to bookmarks and searches - which allows the belongs_to connection to be any model that implements is_blacklight_user

+++ lib/blacklight.rb
@@ -1,3 +1,6 @@
+require 'will_paginate'
+require 'marc'
+
 module Blacklight
 
   autoload :Configurable, 'blacklight/configurable'
@@ -8,11 +11,12 @@ module Blacklight
   
   autoload :SolrHelper, 'blacklight/solr_helper'
   
-
   autoload :Exceptions, 'blacklight/exceptions'
 
   autoload :User, 'blacklight/user'
-  
+
+  autoload :CommaLinkRenderer, 'blacklight/comma_link_renderer'
+
   extend Configurable
   extend SearchFields
   
@@ -36,9 +40,15 @@ module Blacklight
   def self.jruby?
     defined?(RUBY_ENGINE) && RUBY_ENGINE == "jruby" 
   end
+
+  def self.solr_file
+    "#{::Rails.root.to_s}/config/solr.yml"
+  end
   
   def self.init
     
+
+    raise "You are missing a solr configuration file: #{solr_file}. Have you run \"rails generate blacklight\"?" unless File.exists?(solr_file) 
     solr_config = YAML::load(File.open("#{::Rails.root.to_s}/config/solr.yml"))
     raise "The #{::Rails.env} environment settings were not found in the solr.yml config" unless solr_config[::Rails.env]
